requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.9.19
awscrt==0.16.7
Babel==2.9.1
beautifulsoup4==4.12.2
bleach==6.0.0
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.1.0
chevron==0.13.1
cloud-init==22.2.2
cmake==3.26.4
colorama==0.4.4
configobj==5.0.6
contourpy==1.1.0
cryptography==36.0.1
cycler==0.11.0
datasets==2.13.0
dbus-python==1.2.18
dill==0.3.6
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.2
environs==9.5.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gdown==4.7.1
googledrivedownloader==0.4
gpg==1.15.1
grpcio==1.53.0
huggingface-hub==0.15.1
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.13.1
kiwisolver==1.4.4
libcomps==0.1.18
lit==16.0.5.post0
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.19.0
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.10.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.0.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkginfo==1.9.6
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
protobuf==4.23.2
psutil==5.9.5
pyarrow==12.0.1
pycparser==2.20
pydantic==1.10.9
Pygments==2.15.1
pymilvus==2.2.8
pyparsing==3.0.9
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==5.4.1
readme-renderer==37.3
release-notification==1.2
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
safetensors==0.3.1
SecretStorage==3.3.3
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
soupsieve==2.4.1
support-info==1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
towhee==1.1.0
towhee.models==1.1.0
tqdm==4.65.0
triton==2.0.0
twine==4.0.2
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
